% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_adjl_g_e_helper.R
\name{adj_G_E}
\alias{adj_G_E}
\title{adj_G_E}
\usage{
adj_G_E(queried_data)
}
\arguments{
\item{queried_data}{- data queried from KEGG}
}
\value{
vector containing relevant adjacent genes and enzymes
}
\description{
adj_G_E
}
