% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_list_to_igraph.R
\name{adj_list_to_igraph}
\alias{adj_list_to_igraph}
\title{adj_list_to_igraph}
\usage{
adj_list_to_igraph(adjl_G_S)
}
\arguments{
\item{adjl_G_S}{adjacency list containing relevant adjacent SNPs/KEGG genes}
}
\value{
an igraph object, containing a network representing all the KEGG data
}
\description{
internal function that assembles all the KEGG data into a network/graph
}
