% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_adjl_r_e_helper.R
\name{adj_R_E}
\alias{adj_R_E}
\title{Title}
\usage{
adj_R_E(queried_data)
}
\arguments{
\item{queried_data}{}
}
\value{
vector containing relevant adjacent enzymes and reactions
}
\description{
Title
}
