% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_adjl_rp_c_helper.R
\name{adj_RP_C}
\alias{adj_RP_C}
\title{adj_RP_C}
\usage{
adj_RP_C(name)
}
\arguments{
\item{name}{- vector of all KEGG reaction names}
}
\value{
vector containing relevant adjacent reaction pairs and compounds/metabolites
}
\description{
adj_RP_C
}
