% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_adjl_rp_r_helper.R
\name{adj_RP_R}
\alias{adj_RP_R}
\title{adj_RP_R}
\usage{
adj_RP_R(queried_data)
}
\arguments{
\item{queried_data}{- data queried from KEGG}
}
\value{
vector containing relevant adjacent reaction pairs and reactions
}
\description{
adj_RP_R
}
