% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjl_to_G.R
\name{adjl_to_G}
\alias{adjl_to_G}
\title{adj_to_G}
\usage{
adjl_to_G(adjl_G_S)
}
\arguments{
\item{adjl_G_S}{- adjacency list containing relevant adjacencies between inputted SNPs and genes from KEGG}
}
\value{
igraph object representing total IMON for inputted SNPs
}
\description{
Internal function that constructs an IMON (Integrated Multi-Omic Network) for an inputted adjacency list containing adjacency information between KEGG genes and queried SNPs.
}
