% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjl_to_G_grouped.R
\name{adjl_to_G_grouped}
\alias{adjl_to_G_grouped}
\title{adjl_to_G_grouped}
\usage{
adjl_to_G_grouped(
  adjl_G_S,
  unique_group_names,
  unique_group_cols,
  group_snps,
  colour_groups,
  ego,
  progress_bar
)
}
\arguments{
\item{adjl_G_S}{- adjacency list containing relevant adjacencies between inputted SNPs and genes from KEGG}

\item{unique_group_names}{- a list of the unique group/variable names in the provided GWAS Catalog association file}

\item{unique_group_cols}{- a list of unique colours for each unique group/variable in the provided GWAS Catalog association file}

\item{group_snps}{- a recursive list containing the lists of SNPs belonging to each unique group/variable in the provided GWAS Catalog association file}

\item{colour_groups}{- boolean: whether or not user has chosen to colour the network by the unique group/variables in the provided GWAS Catalog association file}

\item{ego}{- the egocentric order (centred around the SNPs in the network) in which to build the network i.e. pathlength from SNPs downwards towards the metabolome}

\item{progress_bar}{- boolean: whether or not user has decided to have a progress bar print to the console}
}
\value{
- an igraph object containing the IMON
}
\description{
Internal function that constructs either a variable-coloured or uncoloured IMON (Integrated Multi-Omic Network) for an inputted adjacency list containing adjacency information between KEGG genes and queried SNPs.
}
