% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_IMON.R
\name{colour_IMON}
\alias{colour_IMON}
\title{colour network by categorical group levels}
\usage{
colour_IMON(G)
}
\arguments{
\item{G}{- igraph object containing uncoloured IMON}
}
\value{
- igraph object containing coloured IMON
}
\description{
colour network by categorical group levels
}
