% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_network_by_groups.R
\name{colour_network_by_groups}
\alias{colour_network_by_groups}
\title{colour IMON nodes and edges by chosen grouping variable}
\usage{
colour_network_by_groups(G)
}
\arguments{
\item{G}{- igraph object containing IMON}
}
\value{
- igraph object containing IMON with coloured nodes/edge attributes by grouping variable
}
\description{
colour IMON nodes and edges by chosen grouping variable
}
