% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSNP_query_check_helper.R
\name{dbSNP_query_check}
\alias{dbSNP_query_check}
\title{dbSNP_query_check}
\usage{
dbSNP_query_check(query)
}
\arguments{
\item{query}{- raw query data from NCBI dbSNP API}
}
\value{
- vector containing either 0 (denoting successful query) or NA (unsuccessful query)
}
\description{
dbSNP_query_check
}
