% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSNP_query_clean_helper.R
\name{dbSNP_query_clean}
\alias{dbSNP_query_clean}
\title{Title}
\usage{
dbSNP_query_clean(query)
}
\arguments{
\item{query}{- dbSNP query}
}
\value{
- dataframe of separate chromosome number, position and ID
}
\description{
Title
}
