% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_IMON.R
\name{decompose_IMON}
\alias{decompose_IMON}
\title{IMON decomposer}
\usage{
decompose_IMON(G)
}
\arguments{
\item{G}{- igraph object containing non-fully connected IMON}
}
\value{
- list of igraph objects containing fully connected IMONs
}
\description{
IMON decomposer
}
