% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ego_IMON.R
\name{ego_IMON}
\alias{ego_IMON}
\title{ego_IMON}
\usage{
ego_IMON(G, ego)
}
\arguments{
\item{G}{- igraph object representing IMON}

\item{ego}{- the selected ego-centred path length}
}
\value{
- ego-centred IMON set at desired path length
}
\description{
ego_IMON
}
