% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ego_IMON_2.R
\name{ego_IMON_2}
\alias{ego_IMON_2}
\title{ego_IMON}
\usage{
ego_IMON_2(G, ego, grouped)
}
\arguments{
\item{G}{- igraph object representing IMON}

\item{ego}{- the selected ego-centred path length}
}
\value{
- ego-centred IMON set at desired path length
}
\description{
internal function for trimming IMON to ego-centred (centred around SNPs) to specified order (pathway length from SNPs)
}
