% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_snp_vector.R
\docType{data}
\name{er_snp_vector}
\alias{er_snp_vector}
\title{Summary-level GWAS data vector for estrogen-receptor positive breast cancer
(EFO_1000649)}
\format{
A vector with 110 elements
}
\usage{
data(er_snp_vector)
}
\description{
A dataset containing a vector of SNPs (summary-level GWAS data) associated
with estrogen-receptor positive breast cancer (EFO_1000649), collated by the
GWAS Catalog.
}
\keyword{datasets}
