% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_cleanup_helper.R
\name{gene_cleanup}
\alias{gene_cleanup}
\title{Title}
\usage{
gene_cleanup(queried_data)
}
\arguments{
\item{queried_data}{}
}
\value{
This function helps to cleans up queried KEGG gene recursive lists + filters by genes with adjacent enzymes
}
\description{
Title
}
