% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_compound_hashmap.R
\name{get_compound_hashmap}
\alias{get_compound_hashmap}
\title{get_compound_hashmap}
\usage{
get_compound_hashmap()
}
\value{
hashmap/recursive list of associated KEGG IDs and compound/metabolite names
}
\description{
This function creates a hashmap of associated KEGG IDs and compound/metabolite names
}
