% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Kegg_Data.R
\name{Get_Kegg_Data}
\alias{Get_Kegg_Data}
\title{Get_Kegg_Data}
\usage{
Get_Kegg_Data(CPU = c(2, 1), sleep = 5)
}
\arguments{
\item{CPU}{The number of cores to use when making KEGGREST API Get requests (default = 2). If CPU > 1, parallel requests will be made.}

\item{sleep}{The amount of sleep (seconds) between a potential caught API error and the next attempt (default = 5)}
}
\value{
Rds files containing adjacency lists for all KEGG network data
}
\description{
This function pulls the latest version of the KEGG database
}
