% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_KEGG_reaction_data.R
\name{get_KEGG_reaction_data}
\alias{get_KEGG_reaction_data}
\title{get_KEGG_reaction_data}
\usage{
get_KEGG_reaction_data(CPU = 1, sleep = 5, progress_bar = c(TRUE, FALSE))
}
\arguments{
\item{CPU}{The number of cores to use when making KEGGREST API Get requests
(default = 2). If CPU > 1, parallel requests will be made. Note, a maximum
of two workers can be used.}

\item{sleep}{The amount of sleep (seconds) between a potential caught API error
and the next attempt (default = 5)}

\item{progress_bar}{Boolean (default = TRUE) argument that controls whether or
not a progress bar for calculations/KEGGREST API GET requests should be printed
to the console}
}
\value{
Rds files containing adjacency lists for all KEGG network data
}
\description{
This function pulls the latest version of the KEGG database
}
\examples{
PanViz::get_KEGG_reaction_data(CPU = 1, sleep = 5, progress_bar = FALSE)
}
