% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_catalog_tsv_to_dataframe.R
\name{GWAS_catalog_tsv_to_dataframe}
\alias{GWAS_catalog_tsv_to_dataframe}
\title{GWAS Catalog tsv to cleaned dataframe}
\usage{
GWAS_catalog_tsv_to_dataframe(file)
}
\arguments{
\item{file}{- The file name of a GWAS Catalog association .tsv file}
}
\value{
A processed dataframe containing only the columns including GWAS studies, traits/phenotypes and relevant SNPs in NCBI standard accession number naming convention
}
\description{
GWAS Catalog tsv to cleaned dataframe
}
\examples{
##getting directory path to GWAS Catalog associtation .tsv file:
path = system.file("extdata", "gwas-association-downloaded_2021-09-13-EFO_1000649.tsv", package="PanViz")
##opening/cleaning data:
df <- GWAS_catalog_tsv_to_dataframe(file = path)

}
