% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_data_reader.R
\name{GWAS_data_reader}
\alias{GWAS_data_reader}
\title{GWAS_data_reader}
\usage{
GWAS_data_reader(file, snp_col, study_col, trait_col)
}
\arguments{
\item{file}{- Character (string) containing the directory path to a .tsv or
.csv file containing summary level GWAS data, typically this can be sourced
from major GWAS databases such as the GWAS Catalog or GWAS Central.}

\item{snp_col}{- Character (string) reflecting the column name containing
the SNP (standard dbSNP accession number, e.g. rs992531) data. In data sourced
from the GWAS Catalog, this column will typically be named "SNPS" and in GWAS
Central this will typically be "Source Marker Accession".}

\item{study_col}{- Character (string) reflecting the column name containing
the study names associated with each SNP. In data sourced from the GWAS Catalog,
this column will typically be named "STUDY" and in GWAS Central this will typically
be "Study Name".}

\item{trait_col}{- Character (string) reflecting the column name containing
the trait/phenotype names associated with each SNP. In data sourced from the
GWAS Catalog, this column will typically be named "DISEASE/TRAIT" and in GWAS
Central this will typically be "Annotation Name".}
}
\value{
A processed dataframe containing only the columns including GWAS studies,
traits/phenotypes and relevant SNPs in NCBI standard accession number naming
convention
}
\description{
GWAS_data_reader
}
\examples{
##getting directory path to GWAS Catalog association .tsv file:
path = system.file("extdata",
  "gwas-association-downloaded_2021-09-13-EFO_1000649.tsv",
  package="PanViz")
##opening/cleaning data:
df <- PanViz::GWAS_data_reader(file = path,
  snp_col = "SNPS",
  study_col = "STUDY",
  trait_col = "DISEASE/TRAIT")
##getting directory path to GWAS Central association .tsv file:
path = system.file("extdata", "GWASCentralMart_ERplusBC.tsv",
  package="PanViz")
##opening/cleaning data:
df <- PanViz::GWAS_data_reader(file = path,
  snp_col = "Source Marker Accession",
  study_col = "Study Name",
  trait_col = "Annotation Name")

}
