% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_hex_col_mix.R
\name{multi_hex_col_mix}
\alias{multi_hex_col_mix}
\title{multi_hex_col_mix}
\usage{
multi_hex_col_mix(col_vector)
}
\arguments{
\item{col_vector}{- vector of hex colours}
}
\value{
- a single mixed hex color from inputted hex codes
}
\description{
This is a helper function that merges any vector of hex colours
}
