% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCBI_clean_helper.R
\name{NCBI_clean}
\alias{NCBI_clean}
\title{NCBI_clean}
\usage{
NCBI_clean(queried_data)
}
\arguments{
\item{queried_data}{- input queried NCBI gene data}
}
\value{
remove genes with no genomic information from NCBI query
}
\description{
NCBI_clean
}
