% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCBI_clean_2_helper.R
\name{NCBI_clean_2}
\alias{NCBI_clean_2}
\title{NCBI_clean_2}
\usage{
NCBI_clean_2(queried_data)
}
\arguments{
\item{queried_data}{- rentrez object queried from NCBI}
}
\value{
return chromosome location, start and end position of gene from NCBI query
}
\description{
NCBI_clean_2
}
