% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCBI_dbSNP_query.R
\name{NCBI_dbSNP_query}
\alias{NCBI_dbSNP_query}
\title{NCBI_dbSNP_query}
\usage{
NCBI_dbSNP_query(snp_list)
}
\arguments{
\item{snp_list}{- list of SNPs to be queried via NCBI dbSNP API}
}
\value{
- raw output from NCBI dbSNP API
}
\description{
NCBI_dbSNP_query
}
