% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCBI_gene_locations.R
\name{NCBI_Gene_Locations}
\alias{NCBI_Gene_Locations}
\title{NCBI Gene Location}
\usage{
NCBI_Gene_Locations(Gene_Enterez_IDs)
}
\arguments{
\item{Gene_Enterez_IDs}{NCBI (enterez) Gene Ids to query}
}
\value{
dataframe containing gene IDs, chromosome number and chromosome start/stop sites
}
\description{
NCBI Gene Location
}
