% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_cleanup_helper.R
\name{reaction_cleanup}
\alias{reaction_cleanup}
\title{reaction_cleanup}
\usage{
reaction_cleanup(queried_data)
}
\arguments{
\item{queried_data}{- input queried KEGG reaction data}
}
\value{
Trimmed recursive lists containing queried KEGG reaction data
}
\description{
This function helps to cleans up queried KEGG reaction recursive lists + separates compound/metabolite and reaction pair data into new sections
}
