% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reactions_Get_All.R
\name{Reactions_Get_All}
\alias{Reactions_Get_All}
\title{Reactions_Get_All}
\usage{
Reactions_Get_All(CPU = c(2, 1), sleep = 5)
}
\arguments{
\item{CPU}{The number of cores to use when making KEGGREST API Get requests (default = 2). If CPU > 1, parallel requests will be made.}

\item{sleep}{The amount of sleep between a potential caught API error and the next attempt (default = 5)}
}
\value{
Rds files for all relevant adjacency lists
}
\description{
This function constructs adjacency lists for compounds, reactions and enzymes listed within the KEGG database
}
