% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry.R
\name{retry}
\alias{retry}
\title{Retry function}
\usage{
retry(
  expr,
  isError = function(x) "try-error" \%in\% class(x),
  maxErrors = 5,
  sleep = 0
)
}
\arguments{
\item{expr}{This is the function you want to catch and handle errors from}

\item{isError}{Define the error}

\item{maxErrors}{The maximum number of errrors it should handle from the function}

\item{sleep}{The amount of sleep between a caught error and the next attempt}
}
\description{
Retry function
}
