% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_base_graph_attributes.R
\name{set_base_graph_attributes}
\alias{set_base_graph_attributes}
\title{set_base_graph_attributes}
\usage{
set_base_graph_attributes(G, colour_groups)
}
\arguments{
\item{G}{igraph object containing KEGG network}

\item{colour_groups}{logical - whether or not user has indicated on colouring
the network by categorical variable i.e. study or trait/phenotype (only available
via PanViz::get_grouped_IMON())}
}
\value{
igraph object with node attributes set
}
\description{
set_base_graph_attributes
}
