% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_snp_grouping.R
\name{set_snp_grouping}
\alias{set_snp_grouping}
\title{snp grouping by chosen categorical variable}
\usage{
set_snp_grouping(G, unique_group_names, unique_group_cols, group_snps)
}
\arguments{
\item{G}{- igraph object containing IMON}

\item{unique_group_names}{- vector containing unique grouping variable names}

\item{unique_group_cols}{- vector containing unique grouping colours for each variable}

\item{group_snps}{- snps split by each variable/group}
}
\value{
- igraph object containing IMON with labelled and coloured snps by grouping variable
}
\description{
snp grouping by chosen categorical variable
}
