% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_gene_chr_match.R
\name{snp_gene_chr_match}
\alias{snp_gene_chr_match}
\title{snp_gene_chr_match}
\usage{
snp_gene_chr_match(snp_loc, gene_loc)
}
\arguments{
\item{snp_loc}{- snp locations}

\item{gene_loc}{- dataframe of genes and their chromosome numbers and start/stop positions}
}
\value{
- a recursive list of gene with their relative snps that have the same chromosome number
}
\description{
snp_gene_chr_match
}
