% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_gene_map.R
\name{snp_gene_map}
\alias{snp_gene_map}
\title{Fast vectorised SNP to gene chromosome number and genomic location mapping}
\usage{
snp_gene_map(gene_loc, snp_loc)
}
\arguments{
\item{gene_loc}{dataframe containing KEGG genes and relevant chromosome number and positions}

\item{snp_loc}{dataframe containing queried SNPs and relevant chromosome number and positions}
}
\value{
an adjacency list of SNPs with their relevant mapped genes to their genomic location
}
\description{
Fast vectorised SNP to gene chromosome number and genomic location mapping
}
