% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_IMON.R
\name{decompose_IMON}
\alias{decompose_IMON}
\title{decompose_IMON}
\usage{
decompose_IMON(G)
}
\arguments{
\item{G}{- igraph object containing non-fully connected IMON}
}
\value{
- list of igraph objects, where each index contains a fully connected
IMON
}
\description{
This function returns a list of fully connected IMONs from a
single parent unconnected IMON.
}
\examples{
data("er_snp_vector")
G <- PanViz::get_IMON(snp_list = er_snp_vector, ego = 5, save_file = FALSE)
G_list <- decompose_IMON(G)

}
