% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_cleanup_helper.R
\name{gene_cleanup}
\alias{gene_cleanup}
\title{gene_cleanup}
\usage{
gene_cleanup(queried_data)
}
\arguments{
\item{queried_data}{queried KEGG gene data (recursive list as returned from KEGGREST)}
}
\value{
cleaned up queried KEGG gene recursive lists, filtered by genes with adjacent enzymes
}
\description{
Internal function for cleaning up queried KEGG gene data
}
\examples{
\dontrun{
Query_Genes_Data <- lapply(Query_Genes_Data, gene_cleanup)
}

}
