% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_compound_hashmap.R
\name{get_compound_hashmap}
\alias{get_compound_hashmap}
\title{Create hashmap of associated KEGG IDs and compound/metabolite names}
\usage{
get_compound_hashmap(CPU = c(2, 1), sleep = 5)
}
\value{
hashmap/recursive list of associated KEGG IDs and compound/metabolite names
}
\description{
Create hashmap of associated KEGG IDs and compound/metabolite names
}
