% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_compound_hashmap.R
\name{get_compound_hashmap}
\alias{get_compound_hashmap}
\title{get_compound_hashmap}
\usage{
get_compound_hashmap(CPU = c(2, 1), sleep = 5)
}
\arguments{
\item{CPU}{The number of cores to use when making KEGGREST API Get requests (default = 2). If CPU > 1, parallel requests will be made.}

\item{sleep}{The amount of sleep (seconds) between a potential caught API error and the next attempt (default = 5)}
}
\value{
hashmap/recursive list of associated KEGG IDs and compound/metabolite names
}
\description{
This function creates a hashmap of associated KEGG IDs and compound/metabolite names
}
